feat: update solana endpoints to match the ones in extension code base (#76)
<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?

Are there any issues or other links reviewers should consult to
understand this pull request better? For instance:

* Fixes #12345
* See: #67890
-->

## **Description**

Make sure the Solana endpoints called by test dapp are the same as the
ones that's are already present in
[privacy-snapshot.json](https://github.com/MetaMask/metamask-extension/blob/main/privacy-snapshot.json)
for metamask-extension, for consistency sake.

* Fixes: https://github.com/MetaMask/MetaMask-planning/issues/5123